Now tests deleting a non existent database

+++ src/test/java/com/codebrane/couchdb/DeleteDBTest.java
@@ -2,6 +2,7 @@ package com.codebrane.couchdb;
 
 import org.junit.Test;
 import org.junit.Assert;
+import static junit.framework.Assert.fail;
 
 /**
  * Test that deletes a CouchDB database
@@ -12,11 +13,33 @@ public class DeleteDBTest extends CouchDBTest {
   @Test
   public void test() {
     System.out.println("DeleteDBTest");
-    CouchPotato cp = new CouchPotato(props.getString(PROP_COUCHDBSERVER));
+    CouchPotato cp = new CouchPotato(couchDBServer);
     cp.connect();
-    CouchPotatoResult cpResult = cp.deleteDatabase(TEST_DB_NAME);
-    Assert.assertNotNull(cpResult);
-    Assert.assertEquals("true", cpResult.ok);
-    cp.disconnect();
+
+    boolean found = true;
+    try {
+      // Try to delete a non existent database
+      cp.deleteDatabase("nonexistentdb");
+    }
+    catch(CouchPotatoException cpe) {
+      Assert.assertEquals(CouchPotatoException.OBJECT_NOT_FOUND, cpe.getReason());
+      found = false;
+    }
+    if (found) {
+      cp.disconnect();
+      fail("Deleted non existent database!");
+    }
+
+    try {
+      CouchPotatoResult cpResult = cp.deleteDatabase(TEST_DB_NAME);
+      Assert.assertNotNull(cpResult);
+      Assert.assertEquals("true", cpResult.ok);
+    }
+    catch(CouchPotatoException cpe) {
+      fail(cpe.getMessage());
+    }
+    finally {
+      cp.disconnect();
+    }
   }
 }
